fedora: Rename directories from fcos to fedora
We can assume throughout these repositories that the target OSes are
actually going to be for bootable containers. That is, that the
VARIANT_ID is "coreos".

To make things simpler, this commit sets the path to "fedora" instead of
"fcos" so it matches the way the other directories (rhel/centos) look
like. This way, the path can be elucidated based on the ${ID} from
/etc/os-release without any further processing.

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>